move files to examples
