+++ provers/stark/src/transcript.rs
@@ -7,11 +7,7 @@ use lambdaworks_math::{
     traits::{ByteConversion, Serializable},
     unsigned_integer::element::U256,
 };
-use miden_core::StarkField;
 use sha3::{Digest, Keccak256};
-
-#[cfg(feature = "winter_compatibility")]
-use miden_core::Felt;
 
 pub trait IsStarkTranscript<F: IsField> {
     fn append_field_element(&mut self, element: &FieldElement<F>);
@@ -151,50 +147,6 @@ impl IsStarkTranscript<Stark252PrimeField> for StoneProverTranscript {
         bytes.copy_from_slice(&self.sample(8));
         let u64_val: u64 = u64::from_be_bytes(bytes);
         u64_val % upper_bound
-    }
-}
-
-pub struct MidenProverTranscript {
-    hasher: Keccak256,
-}
-
-impl MidenProverTranscript {
-    pub fn new(data: &[u8]) -> Self {
-        let mut res = Self {
-            hasher: Keccak256::new(),
-        };
-        res.append_bytes(data);
-        res
-    }
-}
-
-#[cfg(feature = "winter_compatibility")]
-impl IsStarkTranscript<Felt> for MidenProverTranscript {
-    fn append_field_element(&mut self, element: &FieldElement<Felt>) {
-        self.append_bytes(&element.value().to_bytes_be());
-    }
-
-    fn append_bytes(&mut self, new_bytes: &[u8]) {
-        self.hasher.update(&mut new_bytes.clone());
-    }
-
-    fn state(&self) -> [u8; 32] {
-        self.hasher.clone().finalize().into()
-    }
-
-    fn sample_field_element(&mut self) -> FieldElement<Felt> {
-        let mut bytes = self.state()[..8].try_into().unwrap();
-        let mut x = u64::from_be_bytes(bytes);
-        while x >= Felt::MODULUS {
-            self.append_bytes(&bytes);
-            bytes = self.state()[..8].try_into().unwrap();
-            x = u64::from_be_bytes(bytes);
-        }
-        FieldElement::const_from_raw(Felt::new(x))
-    }
-
-    fn sample_u64(&mut self, upper_bound: u64) -> u64 {
-        u64::from_be_bytes(self.state()[..8].try_into().unwrap()) % upper_bound
     }
 }
