+++ winterfell_adapter/src/examples/mod.rs
@@ -1,3 +1,5 @@
 pub mod cubic;
 pub mod fibonacci_2_terms;
 pub mod fibonacci_rap;
+pub mod miden_vm;
+
